# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface BuildingInfo {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  buildingType?: string | null;
  /** @nullable */
  constructionYear?: number | null;
  /** @nullable */
  totalSurface?: number | null;
  /** @nullable */
  heatedSurface?: number | null;
  /** @nullable */
  numberOfFloors?: number | null;
  /** @nullable */
  numberOfOccupants?: number | null;
  /** @nullable */
  climateZone?: string | null;
}

export interface EnergyConsumption {
  /** @nullable */
  totalConsumption?: number | null;
  /** @nullable */
  electricityConsumption?: number | null;
  /** @nullable */
  gasConsumption?: number | null;
  /** @nullable */
  heatingConsumption?: number | null;
  /** @nullable */
  coolingConsumption?: number | null;
  /** @nullable */
  hotWaterConsumption?: number | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  referenceYear?: number | null;
}

export interface EnergyCost {
  /** @nullable */
  totalCost?: number | null;
  /** @nullable */
  electricityCost?: number | null;
  /** @nullable */
  gasCost?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  referenceYear?: number | null;
}

export interface CO2Emissions {
  /** @nullable */
  totalEmissions?: number | null;
  /** @nullable */
  electricityEmissions?: number | null;
  /** @nullable */
  gasEmissions?: number | null;
  /** @nullable */
  unit?: string | null;
}

export interface EnvelopeData {
  /** @nullable */
  wallInsulation?: string | null;
  /** @nullable */
  roofInsulation?: string | null;
  /** @nullable */
  floorInsulation?: string | null;
  /** @nullable */
  windowType?: string | null;
  /** @nullable */
  windowSurface?: number | null;
  /** @nullable */
  airTightness?: number | null;
  /** @nullable */
  thermalBridges?: string | null;
}

export interface HVACSystem {
  /** @nullable */
  heatingSystem?: string | null;
  /** @nullable */
  heatingEfficiency?: number | null;
  /** @nullable */
  coolingSystem?: string | null;
  /** @nullable */
  coolingEfficiency?: number | null;
  /** @nullable */
  ventilationType?: string | null;
  /** @nullable */
  hotWaterSystem?: string | null;
}

export interface EnergyLabel {
  /** @nullable */
  currentLabel?: string | null;
  /** @nullable */
  primaryEnergyConsumption?: number | null;
  /** @nullable */
  referenceConsumption?: number | null;
  /** @nullable */
  energyIndex?: number | null;
}

export type RecommendationPriority =
  (typeof RecommendationPriority)[keyof typeof RecommendationPriority];

export const RecommendationPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Recommendation {
  category: string;
  description: string;
  /** @nullable */
  estimatedSaving?: number | null;
  /** @nullable */
  estimatedCost?: number | null;
  priority: RecommendationPriority;
  /** @nullable */
  paybackPeriod?: number | null;
}

export interface RawExtractedField {
  key: string;
  value: string;
  /** @nullable */
  section?: string | null;
}

export interface AuditReport {
  id: number;
  fileName: string;
  fileType: string;
  uploadedAt: string;
  buildingInfo: BuildingInfo;
  energyConsumption: EnergyConsumption;
  energyCost: EnergyCost;
  co2Emissions: CO2Emissions;
  envelopeData: EnvelopeData;
  hvacSystem: HVACSystem;
  energyLabel: EnergyLabel;
  recommendations: Recommendation[];
  rawFields: RawExtractedField[];
}

export interface AuditReportSummary {
  id: number;
  fileName: string;
  fileType: string;
  uploadedAt: string;
  /** @nullable */
  buildingName?: string | null;
  /** @nullable */
  buildingType?: string | null;
  /** @nullable */
  totalSurface?: number | null;
  /** @nullable */
  currentLabel?: string | null;
  /** @nullable */
  totalConsumption?: number | null;
}

export type AuditStatsLabelDistribution = { [key: string]: number };

export type AuditStatsBuildingTypeDistribution = { [key: string]: number };

export interface AuditStats {
  totalReports: number;
  /** @nullable */
  averageConsumption?: number | null;
  /** @nullable */
  averageSurface?: number | null;
  labelDistribution: AuditStatsLabelDistribution;
  buildingTypeDistribution: AuditStatsBuildingTypeDistribution;
  /** @nullable */
  totalCO2Saved?: number | null;
}

export type UploadAuditFileBody = {
  file: Blob;
};