# lib/api-zod/src/generated/types/auditReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildingInfo } from "./buildingInfo";
import type { CO2Emissions } from "./cO2Emissions";
import type { EnergyConsumption } from "./energyConsumption";
import type { EnergyCost } from "./energyCost";
import type { EnergyLabel } from "./energyLabel";
import type { EnvelopeData } from "./envelopeData";
import type { HVACSystem } from "./hVACSystem";
import type { RawExtractedField } from "./rawExtractedField";
import type { Recommendation } from "./recommendation";

export interface AuditReport {
  id: number;
  fileName: string;
  fileType: string;
  uploadedAt: Date;
  buildingInfo: BuildingInfo;
  energyConsumption: EnergyConsumption;
  energyCost: EnergyCost;
  co2Emissions: CO2Emissions;
  envelopeData: EnvelopeData;
  hvacSystem: HVACSystem;
  energyLabel: EnergyLabel;
  recommendations: Recommendation[];
  rawFields: RawExtractedField[];
}