# lib/api-zod/src/generated/types/auditReportSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditReportSummary {
  id: number;
  fileName: string;
  fileType: string;
  uploadedAt: Date;
  /** @nullable */
  buildingName?: string | null;
  /** @nullable */
  buildingType?: string | null;
  /** @nullable */
  totalSurface?: number | null;
  /** @nullable */
  currentLabel?: string | null;
  /** @nullable */
  totalConsumption?: number | null;
}