# lib/api-zod/src/generated/types/auditStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditStatsBuildingTypeDistribution } from "./auditStatsBuildingTypeDistribution";
import type { AuditStatsLabelDistribution } from "./auditStatsLabelDistribution";

export interface AuditStats {
  totalReports: number;
  /** @nullable */
  averageConsumption?: number | null;
  /** @nullable */
  averageSurface?: number | null;
  labelDistribution: AuditStatsLabelDistribution;
  buildingTypeDistribution: AuditStatsBuildingTypeDistribution;
  /** @nullable */
  totalCO2Saved?: number | null;
}