# lib/api-zod/src/generated/types/cO2Emissions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface CO2Emissions {
  /** @nullable */
  totalEmissions?: number | null;
  /** @nullable */
  electricityEmissions?: number | null;
  /** @nullable */
  gasEmissions?: number | null;
  /** @nullable */
  unit?: string | null;
}