# lib/api-zod/src/generated/types/energyConsumption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyConsumption {
  /** @nullable */
  totalConsumption?: number | null;
  /** @nullable */
  electricityConsumption?: number | null;
  /** @nullable */
  gasConsumption?: number | null;
  /** @nullable */
  heatingConsumption?: number | null;
  /** @nullable */
  coolingConsumption?: number | null;
  /** @nullable */
  hotWaterConsumption?: number | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  referenceYear?: number | null;
}