# lib/api-zod/src/generated/types/energyCost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyCost {
  /** @nullable */
  totalCost?: number | null;
  /** @nullable */
  electricityCost?: number | null;
  /** @nullable */
  gasCost?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  referenceYear?: number | null;
}