# lib/api-zod/src/generated/types/energyLabel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyLabel {
  /** @nullable */
  currentLabel?: string | null;
  /** @nullable */
  primaryEnergyConsumption?: number | null;
  /** @nullable */
  referenceConsumption?: number | null;
  /** @nullable */
  energyIndex?: number | null;
}