# lib/api-zod/src/generated/types/envelopeData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface EnvelopeData {
  /** @nullable */
  wallInsulation?: string | null;
  /** @nullable */
  roofInsulation?: string | null;
  /** @nullable */
  floorInsulation?: string | null;
  /** @nullable */
  windowType?: string | null;
  /** @nullable */
  windowSurface?: number | null;
  /** @nullable */
  airTightness?: number | null;
  /** @nullable */
  thermalBridges?: string | null;
}