# lib/api-zod/src/generated/types/hVACSystem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface HVACSystem {
  /** @nullable */
  heatingSystem?: string | null;
  /** @nullable */
  heatingEfficiency?: number | null;
  /** @nullable */
  coolingSystem?: string | null;
  /** @nullable */
  coolingEfficiency?: number | null;
  /** @nullable */
  ventilationType?: string | null;
  /** @nullable */
  hotWaterSystem?: string | null;
}