# lib/api-zod/src/generated/types/rawExtractedField.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */

export interface RawExtractedField {
  key: string;
  value: string;
  /** @nullable */
  section?: string | null;
}