# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Energy audit report generation API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationPriority } from "./recommendationPriority";

export interface Recommendation {
  category: string;
  description: string;
  /** @nullable */
  estimatedSaving?: number | null;
  /** @nullable */
  estimatedCost?: number | null;
  priority: RecommendationPriority;
  /** @nullable */
  paybackPeriod?: number | null;
}